backend/app/requirements.prod.txt
django
django-extensions
djangorestframework

gunicorn